% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_significance}
\alias{check_significance}
\title{Gets t/z value out of model_emp and compares it to a ctritical value}
\usage{
check_significance(model, critical_value)
}
\arguments{
\item{model}{mixed model you want to check}

\item{critical_value}{t/z value you want to check against}
}
\value{
logical vector
}
\description{
\code{check_significance()} checks the t/z value of all fixed effects and
compares them to the specified critical value. Returns a logical vector
indicating whether one effect is significant or not. TRUE == significant,
FALSE == not significant.
}
