% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{fullpower}
\alias{fullpower}
\title{Function to prepare a mixed model with varied fixed effect estimantes}
\usage{
fullpower(model_emp, data_emp, subvar, fixed_effects, critical_value,
  sample_sizes, n_sim, databased = T, safeguard = T, rnorm = T)
}
\arguments{
\item{data_emp}{pilot data that fits the mixed model of interest}

\item{subvar}{name of the variable that contains the subject´s number
in data_emp}

\item{critical_value}{z/t value to test if a given fixed effect
is significant}

\item{model}{mixed model of interest}
}
\value{
A modified mixed model
}
\description{
\code{prepare_rnorm_model()} changes the fixed effect estimates of the
specified model. It randomly selects "new" fixed effect estimates out of
a normal distribution around the "old" estimates. Non significant effects
in the given model are ignored and no rnorm-value is computet for them.
}
