\name{generateDesignsForManuscript}
\alias{generateDesignsForManuscript}
\title{generateDesignsForManuscript}
\usage{
generateDesignsForManuscript()
}
\value{
list of pairs of design.glmmFG and glh objects
}
\description{
Generate the list of test cases (design.glmmFG and glh
objects) for the manuscript:\cr Kreidler, S. M., Muller, K.
E., & Glueck, D. H. A Power Approximation for Longitudinal
Studies Using the Kenward and Roger Wald Test in the Linear
Mixed Model, In review.
}
\keyword{internal}

