% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_depvar}
\alias{get_depvar}
\title{Gets dependent variable of a mixed model}
\usage{
get_depvar(model)
}
\arguments{
\item{model}{mixed model you want to know the dependent variable of}
}
\description{
\code{get_depvar()} checks the formula of a given mixed model and return
the variable name of the dependent variable.
}
