% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_n}
\alias{get_n}
\title{Gets number of participants in a dataset}
\usage{
get_n(data, simvar)
}
\arguments{
\item{data}{data set you want to get n for}

\item{simvar}{name of variable in data that contains the subject number}
}
\description{
\code{get_n()} checks the column indicating the subject number and returns
the number of unique occurences.
}
