% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{keep_balance}
\alias{keep_balance}
\title{keeps balance of groups in designs with between-subject variables}
\usage{
keep_balance(final_data, simvar, fixed_effects, n_want)
}
\arguments{
\item{final_data}{simulated data}

\item{simvar}{variable that is changed in simulation process}

\item{fixed_effects}{fixed effects in model used to simulate data}

\item{n_want}{levels of simvar the data should have}
}
\description{
keeps balance of groups in designs with between-subject variables
}
