% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{mixedpower}
\alias{mixedpower}
\title{Function running simulation for one varying random effect}
\usage{
mixedpower(
  model,
  data,
  fixed_effects,
  simvar,
  steps,
  critical_value,
  n_sim = 1000,
  SESOI = F,
  databased = T
)
}
\arguments{
\item{model}{lme4 model: mixed model of interest}

\item{data}{data frame: data used to fit the mixed model of interest}

\item{fixed_effects}{vector of character elements: names of variables that
are used as fixed effects in the used model}

\item{simvar}{character element: name of the variable that contains the
random effect we want to simulate along (e.g. subject number or stimuli number)
in the used data frame}

\item{steps}{vector of integers: steps of "simvar" you want to test power
of (e.g. different sample sizes)}

\item{critical_value}{integer: z/t value to test if a given fixed effect
is significant. This can be a single value or a vector containing specific
critical values for each effect}

\item{n_sim}{integer: number of simulations to run}

\item{SESOI}{vector with floats´indicating the desired SESOIs.
If FALSE, no SESOI simulation is run.}

\item{databased}{logical value: indicates whether databased power simulation
shoul be run}
}
\value{
A modified mixed model
}
\description{
\code{mixedpower()} combines a databased and SESOI (smallest effect of interest)
simulation for one specified random effect
}
