% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{mixedpower1}
\alias{mixedpower1}
\title{Function thatr runs the whole mixedpower power process}
\usage{
mixedpower1(model, data, fixed_effects, simvar, steps, critical_value,
  n_sim = 1000, confidence_level = 0.68, databased = T,
  safeguard = F, rnorm = F)
}
\arguments{
\item{model}{lme4 model: mixed model of interest}

\item{data}{data frame: pilot data that fits the mixed model of interest}

\item{fixed_effects}{vector of character elements: names of variables that
 are used as fixed effects in
model emp}

\item{simvar}{charackter element: name of the variable that contains the
random effect we want to simulate along (e.g. subject number or stimuli number)
in data}

\item{steps}{vector of integers: steps for "simvar" you want to test power
of}

\item{critical_value}{integer: z/t value to test if a given fixed effect
is significant}

\item{n_sim}{integer: number of simulations to run}

\item{confidence_level}{float: value between 0-1 indicating the width of the
confidence interval used for the safeguard option}

\item{databased}{logical value: indicates whether databased power simulation
shoul be run}

\item{safeguard}{logical value: indicates whether safeguard power simulation
shoul be run}

\item{rnorm}{logical value: indicates whether rnorm power simulation
shoul be run}
}
\value{
A modified mixed model
}
\description{
\code{mixedpower()} combines all three power options (databased, safeguard,
rnorm). It runs them all by default and returns a data frame with results
for all three options.
}
