% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function from Cookbook-R}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects passed to the function}

\item{plotlist}{List of ggplot objects that can be passed to the function}

\item{cols}{Number of collumns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\description{
\code{multiplot(..., plotlist=NULL, file, cols=1, layout=NULL)} puts
 multiple graphs on one page.
}
