% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{multiplotPower}
\alias{multiplotPower}
\title{Multiple plot function to plot the mmpower output}
\usage{
multiplotPower(output_data, ppi = 300)
}
\arguments{
\item{output_data}{data frame returned from \code{fullpower} function}

\item{ppi}{resolution of plot. default is 300}
}
\description{
\code{multiplotPower(output_data, ppi = 300)} plots the output for every
effect in the mixed model used for power simulation.A png file with all
plots is saved to your working directory.
}
