% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{power_simulation}
\alias{power_simulation}
\title{Function that runs the power simulation}
\usage{
power_simulation(model_emp, data_emp, subvar, fixed_effects, critical_value,
  sample_sizes, n_sim, safeguard = F, rnorm = F)
}
\arguments{
\item{model_emp}{lme4 model: mixed model of interest}

\item{data_emp}{data frame: pilot data that fits the mixed model of interest}

\item{subvar}{charackter element: name of the variable that contains the
subject´s number
in data_emp}

\item{fixed_effects}{vector of character elements: names of variables that
 are used as fixed effects in
model emp}

\item{critical_value}{integer: z/t value to test if a given fixed effect
is significant}

\item{n_sim}{integer: number of simulations to run}

\item{safeguard}{logical value: indicates whether safeguard power simulation
shoul be run}

\item{rnorm}{logical value: indicates whether rnorm power simulation
shoul be run}

\item{sampe_sizes}{vector of integers: sample sizes you want to test power
of}
}
\value{
A modified mixed model
}
\description{
\code{power_simulation()} runs the actual simulation process for either the
databased, safeguard or rnorm power option. Returns a data frame with results
for all fixed effects and all tested sample sizes.
}
