% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepare_safeguard_model}
\alias{prepare_safeguard_model}
\title{Function that computes confidence intervals for fixed effects and
sets them to the lower boundary of this CI}
\usage{
prepare_safeguard_model(model, confidence_level, critical_value)
}
\arguments{
\item{model}{mixed model of interest}

\item{confidence_level}{level of confidence that the parameter lies
in the interval}

\item{critical_value}{z/t value to test if a given fixed effect
is significant}
}
\value{
A modified mixed model
}
\description{
\code{prepare_safeguard_model()} changes the fixed effect estimates of the
specified model. It computes confidence Intervals with the "Wald"- method
and sets the fixed effect estimates to the lower boundary of this confidence
interval. Non significant effects in the given model are ignored and
no safeguard-value is computet for them.
}
