% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_functions.R
\name{R2power}
\alias{R2power}
\title{Function running simulation for two varying random effect}
\usage{
R2power(
  model,
  data,
  fixed_effects,
  simvar,
  steps,
  R2var,
  R2level,
  critical_value,
  n_sim = 1000,
  SESOI = F,
  databased = T,
  maxCores = NULL
)
}
\arguments{
\item{model}{lme4 model: mixed model of interest}

\item{data}{data frame: data used to fit the mixed model of interest}

\item{fixed_effects}{vector of character elements: names of variables that
are used as fixed effects in the used model}

\item{simvar}{character element: name of the variable that contains the
random effect we want to simulate along (e.g. subject number or stimuli number)
in data}

\item{steps}{vector of integers: steps for "simvar" you want to test power
of}

\item{R2var}{character: name of second random effect we want to vary}

\item{R2level}{integer: number of levels for R2var. Right now, the second
random effect can only be changed to a fixed value and not be varied like
simvar}

\item{critical_value}{integer: z/t value to test if a given fixed effect
is significant}

\item{n_sim}{integer: number of simulations to run}

\item{SESOI}{vector with floats´indicating the desired SESOIs.
If FALSE, no SESOI simulation is run.}

\item{databased}{logical value: indicates whether databased power simulation
#' @param maxCores allows to specify the maximum number of cores to be used in
simulation process. If unspecified, R2power will run simulations on all but one cores.}
}
\value{
A modified mixed model
}
\description{
\code{mixedpower()} combines a databased and SESOI (smallest effect of interest)
simulation for a combination of two specified randomeffects
}
