% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateDataset}
\alias{simulateDataset}
\title{Simulate a new data set}
\usage{
simulateDataset(n_want, data, model, simvar, fixed_effects, use_u = F)
}
\arguments{
\item{n_want}{integer: how many subjects should the new data set include?}

\item{data}{data frame: pilot data that fits the mixed model of interest}

\item{model}{lme4 model: mixed model of interest}

\item{simvar}{character element: name of the varaible containing the subject}

\item{fixed_effects}{fixed effects specified in model
number in data}
}
\value{
A modified mixed model
}
\description{
\code{simulateDataset()} builds a new data set with a specified number of
subjects. It uses the \code{lme4::simulate()} function to create new response
values based on the mixed model fittet to the pilot data.
}
