% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateSamplesize}
\alias{simulateSamplesize}
\title{Simulate a new data set}
\usage{
simulateSamplesize(n_want, data_emp, model_emp, subvar)
}
\arguments{
\item{n_want}{integer: how many subjects should the new data set include?}

\item{data_emp}{data frame: pilot data that fits the mixed model of interest}

\item{model_emp}{lme4 model: mixed model of interest}

\item{subvar}{character element: name of the varaible containing the subject
number in data_emp}
}
\value{
A modified mixed model
}
\description{
\code{simulateSamplesize()} builds a new data set with a specified number of
subjects. It uses the \code{lme4::simulate()} function to create new response
values based on the mixed model fittet to the pilot data.
}
