% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepare_rnorm_model}
\alias{prepare_rnorm_model}
\title{Function to prepare a mixed model with varied fixed effect estimantes}
\usage{
prepare_rnorm_model(model, data, simvar, critical_value)
}
\arguments{
\item{model}{mixed model of interest}

\item{data}{pilot data that fits the mixed model of interest}

\item{simvar}{name of the variable that contains the subject??s number
in data}

\item{critical_value}{z/t value to test if a given fixed effect
is significant}
}
\value{
A modified mixed model
}
\description{
\code{prepare_rnorm_model()} changes the fixed effect estimates of the
specified model. It randomly selects "new" fixed effect estimates out of
a normal distribution around the "old" estimates. Non significant effects
in the given model are ignored and no rnorm-value is computet for them.
}
