% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{reset_contrasts}
\alias{reset_contrasts}
\title{Resets specified contrasts in the simulated dataset}
\usage{
reset_contrasts(simulated_data, data, model, fixed_effects)
}
\arguments{
\item{simulated_data}{simulated data set (= output of simulate_samplesize)}

\item{data}{pilot data used for power simulation}

\item{model}{mixed model used for power simulation.
Needs to fit data.}

\item{fixed_effects}{names of variables in data that are used as
fixed effects in model}
}
\value{
Date frame
}
\description{
\code{reset_contrasts()} makes sure that contrasts in the simulated
dataset equals the contrasts in the pilotdata used for simulation.
Because the simulate_sanplesize function loses specified contrasts,
they need to be re-specified.
\code{reset_contrasts()} checks all contrasts on the fixed-effect-variables
in the pilot dataset and copies them to the simulated dataset.
}
