% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateModel}
\alias{simulateModel}
\title{Simulate a new data set}
\usage{
simulateModel(model, data, n_want, simvar, nsim = 100)
}
\arguments{
\item{model}{lme4 model: starting mixed model used for simulation}

\item{data}{data frame: data used to inform the simulation}

\item{n_want}{integer: how many levels should the new model be based on?}

\item{simvar}{character element: name of the varaible containing
the random effect levels}
}
\value{
A modified mixed model
}
\description{
\code{simulateModel()} builds a new model set with a specified number of
levels of a specified random effect. New data sets are simulated and
averaged to return the average model for a specific sample size
}
